feat(frontend): port Frontend::LaunchGameSession + GameEngine::Shutdown
Frontend::LaunchGameSession (0x004EC540) is the funnel from every dialog
launch path into GameSession::Run. Ported as a stdcall full-replacement
hook in wa::frontend::launch_game_session, including pre-launch teardown
(audio + display + frontend hide) and post-game restore (framebuffer
rebuild + audio + dialog re-show).

GameEngine::Shutdown (0x0056DCD0) is now also a Rust port in
engine::hardware_init::shutdown — only WA-side caller was GameSession::Run
(itself fully replaced), so the WA address is trapped.

Add wa_call helpers for the audio/mouse usercall patterns
(call_usercall_eax_esi/_esi/_edi/_edi_stack2). Add MFC wrappers
(CWnd::ShowWindow/MoveWindow/SetFocus/FromHandle, AfxGetModuleState).

Rename DDGame -> GameWorld in code and Ghidra: g_GameWorldInstance,
GameWorldRenderChildren_Maybe, CONSTRUCT_GAME_WORLD, CONSTRUCT_GAME_RUNTIME.
"DDGame" was WA's old codename for the same object. Also rename
MouseModeEventLog_Maybe -> Frontend__OnGraphicsInitError_Maybe (it's a
graphics-init error MessageBox + reset prompt, not a log function).

Skip the alt-display "different DDGame instance" branch in LaunchGameSession's
post-game block — stock WA's path is unconditionally fatal in practice
(g_GameWorldInstance is never written, so the `param_1 != it` guard always
fires; the called Frontend::OnGraphicsInitError dereferences param_1 as a
GLibError* that's actually the GameRuntime vtable, leading to a
Localization::GetString AV that stock WA's outer __try swallows). Skipping
the buggy branch retains the legitimate framebuffer rebuild.

Add Rust-native SEH helper at openwa-game::seh::try_seh — a Vectored
Exception Handler + setjmp/longjmp pair (in naked_asm) that catches AVs
during a closure body and returns Err(()) instead of crashing. Used to
wrap the DSSound destructor's COM-release sequence (matches stock WA's
__try/__except (EXCEPTION_EXECUTE_HANDLER) at 0x00573DD0).

Streaming-audio teardown race: timer was created with TIME_PERIODIC only;
add TIME_KILL_SYNCHRONOUS (0x100) so timeKillEvent waits for any in-flight
callback before returning, eliminating use-after-free during shutdown.

Add 0x68 (PUSH imm32) to startup_checks valid prologues so AfxGetModuleState's
wrapper-style entry at 0x005C83F1 doesn't trip the load-time sanity check.

+++ crates/openwa-game/src/input/keyboard.rs
@@ -211,7 +211,7 @@ pub unsafe extern "thiscall" fn keyboard_alert_user(
 
         let foreground = GetForegroundWindow();
         let session = get_game_session();
-        if !session.is_null() && foreground as u32 == (*session).hwnd {
+        if !session.is_null() && foreground == (*session).hwnd {
             return;
         }
         let frontend_hwnd = *(rb(va::G_FRONTEND_HWND) as *const HWND);
